gen/requirements.txt
markdown-it-py == 2.0.1
precisely == 0.1.9
requests == 2.27.1
